use infinite idleTimeout in the API server
Fixes #833.

The timeout is now a parameter when running the server, which is hard-coded to infinity for real life, and set to 1 second in tests.

This required switching to the Blaze backend (probably a good thing anyway), because of a bug in the Jetty backend that we were using (see http4s/http4s#336). That required a workaround for the bug in the Blaze backend that was the reason we went with the Jetty backend in the first place (see http4s/http4s#337).

Disabled our `FailSafe` middleware, which was only necessary due to another bug in the Jetty backend (see http4s/http4s#338).

Moved the http4s dependencies to the web sub-project, and bumped the version to 0.8.3 for good measure.