fix exception in compiler with un-annotated Expr
Fixes #815.

We were trying to compile an Expr that wasn't part of the original AST,
so had no annotations. This is the easy fix: just use the LogicalPlan
equivalent directly.

Really, `attr` should be partial, so it could fail if you ask for the
annotation for some unknown node. As it is, it has to be unsafe. But
really we'd rather use fixplate probably, so I'm not going there right
now.