Don't push $$ROOT as a rule.
Fixes a case where we would invariably push `$$ROOT` if there was a single simple field alongside an aggregate function (fixes #831). This is a one-line fix in `objectConcat`.

Capture several steps within the Planner as PhaseResults to aid debugging:
- `Logical Plan (projections preferred)`: the rewritten LP actually used in the planner
- `Workflow Builder`: speaks for itself
- `Workflow (raw)`: before any optimizations are applied
- `Workflow (finished)`: after reordering ops and deleting unused fields, but before "crystallization" (see below)

The PhaseResult logged as `Physical Plan` now better represents what is actually executed, because it shows inlining and shape-adjustments that are done in `crystallize`.

Rename `Workflow.finalize` to `crystallize` to avoid clashing with the method defined on every Java class/Scala trait, and wrap the result in a type, `Crystallized`, to avoid errors that resulted from calling it more than once. Thanks to @sellout for the name, which is meant to suggest both a transformation and a kind of "freezing", since this is necessarily the last transformation.

Planner tests now compare against the crystallized plan, which captures more of the process but is slightly noisier in a few cases. In particular, inlining of flattening into $SimpleMap is now verified.