mark field-order test as pending
This query used to push $$ROOT, and that indirectly made it project
both fields at the same time, working around #598. Now it runs more
efficiently, but falls afoul of that MongoDB bug. Since it was only
meant to be a test of that problem, I put it back to pending and
updated the comment to reference the right issue (#598).