In the API, test connection to new mount before saving
Fixes #863.

Call `Backend.test` from the API before saving a new mount (previously it was used only by the now defunct admin UI).

Make that test somewhat more robust by also writing a temp file so that it fails in a read-only configuraiton.

Capture additional errors from MongoDB in our error types. Unfortunately, much of the interesting information is available only in the message strings of the MongoDB exceptions, so for now that's just captured and wrapped.

Also, went back to the simpler server setup for the bulk of the API tests. The server restarting code path is now used only by the one test the depends on it (and tests it), while the rest use the simpler, single-threaded setup. Hopefully this will resolve the flakiness we've seen in Jenkins and elsewhere.