Project a single value for any expression appearing in the group keys
Fixes #459.

New Func `Arbitrary`, to select a single value from a set. Translated to $first for MongoDB.

During compilation, keep track of the compiled form of each grouping key, and then when any of them occurs in a projection (even embedded in a larger expression), wrap with `Arbitrary(...)` to simulate SQL semantics.

Handle an additional case in objectConcat that arises from the new shape.

Update query generator to avoid generating anything that would (correctly) lead to a plan with two consecutive group ops, because that pattern is flagged as a common characteristic of busted plans.

Update integration test queries to remove DISTINCT where it's no longer needed.