Put Expression on the right of various disjunctions.
Expression tends to be the “default” state, and so we’ve been doing
things like `.map(_.swap)` and `.fold(Some(_), None)`. This eliminates
that extra hoop to jump through.

The three types changed are `WorkflowBuilder.Expr`,
`WorkflowBuilder.GroupValue`, and `Reshape.Shape`.

Also, update Scala to 2.11.7 (some dependency was forcing it anyway) and
add a missing `SpliceBuilder` case to `WorkflowBuilder.objectConcat`.