`distinct` without `Workflow`.
SD-949 #done

This reimplements `distinctBy` (and adds `distinct`) without converting
to `Workflow`. It also fixes some test cases that I _think_ should
already have been working, but were overlooked when #283 was fixed.

Most significantly, the required replacing `DocVar` with `Base` for rewriting.

`Base` is a richer abstraction that allows us to distinguish between an
entire document (IE, DocVar.ROOT) and a subset of a document (which was
not distinguished from ROOT previously).

This required various other fixes to work properly:

* add a NEL `apply` to `BsonField`,
* render BSON as JS rather than `repr`,
* move all of the `ExprBuilder` coalescing from `expr` to `normalize`,
* extract `findKeys` from `distinct`,
* move a couple `merge` cases that we preventing splices from unifying
  properly,
* use `apoM` instead of `topDownTransformM` in `normalize` (this prevents
  double-processing (and thus failure) in the case when a substitution
  contains a projection itself),
* add `anaM` and `apoM` recursion schemes,
* convert `Schema` elements from `String` to `BsonField.Name`,
* eliminate an unused `fold1Builders` call, and
* propagate the correct `Schema` from `GroupBuilder`s.