update to http4s 0.8.6
Fixes #886 (now known as SD-888), a stack overflow when reading several thousand rows via the API with gzip encoding.

The bug was fixed in http4s, which is updated here. This commit adds integration tests of the backend and of the API implementation, showing that each is capable of emitting a 100k rows result set.