make static file serving optional and paramterized
SD-959 #done

Adds a new parameter, `-L`, when launching the API server, which specifies the URL path static files should be served from. This will be used to configure the path to the front-end's files, without needing any knowledge hard-coded in the server.

Adds a new page at `/welcome`, to which `/` is redirected if no static files are being served. This will be a default landing page for users starting the bare server, and would be a good place to host some simple tips about how to get started, or links to documentation, etc. For now, it's just a tiny page showing the version.

Fixed an infinite HTTP redirect loop when static files aren't present.