Fix demo app
The importTestData script assumes the test data is in the test source folder of the integration test project. That was changed in a recent commit and the importTestData was not updated to reflect that.

Also moved two new tests that were now in the wrong place because of an automated merge.