Update issues referenced in code.
There are a number of places in the code (pending tests, comments, etc.)
where we refer to issue numbers. They now refer to JIRA issues rather
than the old GitHub issues.

NB: There were three references that I couldn’t find the ticket for. I
left the GitHub numbers there in the hopes that it may help someone in
future.