Introduce FunctorT and TraverseT.
This makes the tests more robust to compiler changes. On failure, we
display both the normalized diff and the original (un-normalized)
result.

`TraverseT` (and `FunctorT`) is a generalization of
`Recursive with Corecursive` providing `traverse` (and `map`) for types
that can’t implement one of `Recursive` or `Corecursive` (EG, `Free` and `Cofree`).

Transformations have been moved from the `Recursive` type class to the
new type classes, and some existing algebras have been generalized from
`Fix` to `T[_[_]]`.