interpret views
Completes SD-977, in terms of what the implementation is capable of, but currently views cannot be added through the API and are ignored if found in the configuration file, due to bugs that make them confusingly broken (see SD-1101).

A new Backend implementation, which interprets all requests involving views by transforming them into queries where the view queries are sub-selects. This means embedding ASTs inside each other, which works surprisingly well but did expose some ways in which the AST is not composable.

Fix ScopeTables to treat sub-selects independently, so that it's not always necessary to qualify all field references when a sub-select is present. This relates to the more subtle problem that's now tracked in SD-1101.

Slight improvements to RenderTree for AnnotatedTree and Provenance annotations in particular.