Reorganize instances in recursionschemes
Tidys the package object a bit and increases the chance of
implicit search finding instances without imports by moving instances
for `Fix` to its companion and the dervied instances for FunctorT and
TraverseT to their companions.

However, had to add imports in a couple of places (and deleted some in
others) but I think it is a net gain for organization.