fix handling of views containing absolute paths
SD-1165 #done

When the ViewBackend is created, it normalizes all paths in every view
query so that they can be composed with each other. Previously, this
normalization ignored absolute paths, leaving us with a mixture of
absolute and relative paths, and if an absolute path happened to refer
to another view, it would not be properly expanded when it came time to
evaluate the (outer) view's query.

Now all paths in view queries are converted to absolute, just to make
it easier to understand/explain. The locations of the views themselves
are still stored as relative and interpreted relative to '/', just
because that's how the Backend trait works. That means the fix here is
really just to convert paths in queries to relative when doing the view
lookup.