handle variables, offset, and limit in GET /query
SD-1150 #done

`GET /query` requests now accept `offset`, `limit`, and `var.` parameters, and `POST` now expects variables to have the `var.` prefix as well.

**Note:** this is a breaking API change. Clients need to be updated to prefix variable names when calling `POST /query`.

Also implemented variables in `/compile` (it was already mentioned in the README, but apparently never implemented).

Tests for all of the above that actually inspect the resulting plans.