handle variable bindings in views
SD-1143 #done

Accept variable bindings in the connectionUri for a view mount. These are parsed out to a Variables map in ViewConfig (so they can be re-serialied later), and substituted into the view query when it is mounted. Unbound variables lead to errors 1) when a view mount is added through the API and 2) whenever a view query is evaluated through Backend's methods.

A few new bits and pieces:
- Yet another error wrapper to embed CompilationError in EnvironmentError.
- A convenience `substVars()` which doesn't need type annotations, since it's now being used in a third place.
- Replace `rewriteRelations()` with `rewriteRelationsM()`.

Add a dependency on http4s-core in the core project, since we're now using http4s's Uri type to parse/render view connectionUris. This should be more robust than the regex that was previously used. Define a settingKey for each dependency version, so the setting can be defined in the main build.sbt and referred to elsewhere.