Rewrite InnerJoin/Filter sequences, putting conditions before, in, or after the join
Completes SD-1063 and SD-910.

The idea is to identify all filtering associated with an inner join in the source plan, breaking down all components into four categories:
- conditions referring to only the left or only the right source, which can be moved ahead of the join (SD-910)
- conditions referring to both sources using `=`, which can be rewritten into the join (SD-1063)
- all others, which have to be done after the join

This is implemented in a new rewriteCrossJoins fold, which uses boundParaM (which is new). Because the new fold introduces new temp names which made it difficult for tests to compare plans, names are now rewritten to a compact sequence starting with `__tmp0` after rewriting cross joins. Nested lets are re-ordered to a consistent sequence. Also, the plan has to be simplified prior to the new fold, so all these steps are now wrapped up in `Optimizer.optimize`, which makes sure to run things in the proper sequence. This replaces the fold with `simplifyƒ` that was being called from several places before.

Added test for two queries which "worked" before but are now massively more efficient.