Allow the version check to succeed when databases cannot be listed
When running with the WiredTiger storage engine and no auth, it seems that listing the available databases silently fails, which was causing the version check to fail during integration tests. With this commit, we simply do that check against the "admin" database, which works whether or not that db actually exists yet.