Add some new refolds.
This adds `ghyloM`, `dyna`, `codyna`, `codynaM` refolds. As well as a
`ganaM` unfold.

`FunctorT.translate` has been removed, as it is identical to
`FunctorT.transAna` (and identical to `.transCata`, since the traversal
order is irrelevant) and attempting to add `TraverseT.translateM` made it clear
that sometimes you do care about whether its bottom-up or
top-down (because the constraints on `F` and `G` differ), so now always
explicitly use `transAna` or `transCata`.

Fixed the type parameter order for `IdOps.ghylo`.

Added a `Nat` type to `matryoshka.fixedpoint`.

Added type aliases for transforms, and a conversion from Transform to
Algebra.

Added optics for algebras and folds.

Made `freeEqual` available outside of tests.