Updates to the build.
Primary motivations were an updated kind-projector and an
easy-to-use console (i.e. one which didn't give a bunch of
inappropriate options to scalac.)

Also updated simulacrum and made necessary source changes.
Opened an issue at Monocle to publish new artifacts, because
that's the one holding back an across-the-board upgrade.

  https://github.com/julien-truffaut/Monocle/issues/399

And updated almost every plugin.