Greatly reduce the general recursion used.
When the next version of WartRemover is released, this will be enforced, but it
can be checked with a snapshot version for the time being. This reduces 40+
instances of recursion to now only 6 (labeled) instances. And this is after a
refactoring of the library already eliminated at least 20 other instances.

Transforms of unfolds (`transAna`, etc.) have moved to `Corecursive`, because
that's the only way to make it possible for `Birecursive` to override them
properly. This required a new `implicit class RecursiveOps` to hold the
enrichments of those unfolds.

Generalized a bunch of things that previously specified types like `Nat` to use
`[N](implicit N: Recursive.Aux[N, Option])`.

Renamed `NelF` to `AndMaybe`. And added fixed-point representations of
NonEmptyList.

`repeatedly`, which truly uses general recursion now correctly labels it with
the `Partial` monad.

Fixes #24, fixes #59, which improve the README.